requirements.txt
dropbox==12.0.2
Flask==3.1.0
fpdf==1.7.2
Pillow==11.1.0
pyfiglet==1.0.2
pyttsx3==2.98
qrcode==8.0
SpeechRecognition==3.14.1
Werkzeug==3.1.3
